Final Push - Cleanup
Changed time to display from Seconds Minutes

Fixed some final collider issues and level spawn issues.

Good for me to call this done for now.

Can add music and shoot level if I really want to but going to just call it. :)

This was a really good learning experience and probs publish to somewhere like Kongregate to disappear into the aether :+1: ^_^

7/10 learning experience